+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 This is an alpha version! The changes listed here are not final.
 
+### Enhancements
+- WordPress.com Reader: add new option to show a link to the WordPress.com Reader in the admin bar.
+
 ### Other changes <!-- Non-user-facing changes go here. This section will not be copied to readme.txt. -->
 - VideoPress: Improve attachment details display with proper i18n support and WordPress core styling.